Revert "renames log module to avoid name conflicts"
This reverts commit 45d9cca24d02a8ae202c27fedb372c350a3188bd.
