+++ services/web/server/tests/unit/with_dbs/01/test_studies_dispatcher_projects.py
@@ -14,7 +14,7 @@ from pytest_simcore.helpers.utils_login import NewUser
 from pytest_simcore.helpers.utils_projects import delete_all_projects
 from pytest_simcore.helpers.utils_services import list_fake_file_consumers
 from simcore_service_webserver.groups_api import auto_add_user_to_groups
-from simcore_service_webserver.logs import setup_logging
+from simcore_service_webserver.log import setup_logging
 from simcore_service_webserver.projects.projects_api import get_project_for_user
 from simcore_service_webserver.studies_dispatcher._projects import (
     UserInfo,
